+++ Commands/Model/GraphToken.cs
@@ -1,17 +1,12 @@
 ﻿using Microsoft.Identity.Client;
+using OfficeDevPnP.Core;
 using PnP.PowerShell.CmdletHelpAttributes;
 using PnP.PowerShell.Commands.Base;
 using PnP.PowerShell.Commands.Utilities;
 using System;
-using System.Collections.Generic;
-using System.IO;
 using System.Linq;
-using System.Management.Automation;
-using System.Runtime.CompilerServices;
 using System.Security;
 using System.Security.Cryptography.X509Certificates;
-using System.Threading;
-using System.Threading.Tasks;
 
 namespace PnP.PowerShell.Commands.Model
 {
@@ -41,9 +36,10 @@ namespace PnP.PowerShell.Commands.Model
         /// <param name="clientId">ClientId to use to acquire the token. Required.</param>
         /// <param name="certificate">Certificate to use to acquire the token. Required.</param>
         /// <returns><see cref="GraphToken"/> instance with the token</returns>
-        public static GraphToken AcquireApplicationToken(string tenant, string clientId, X509Certificate2 certificate)
+        public static GraphToken AcquireApplicationToken(string tenant, string clientId, X509Certificate2 certificate, AzureEnvironment azureEnvironment)
         {
-            return new GraphToken(GenericToken.AcquireApplicationToken(tenant, clientId, $"{BaseAuthority}{tenant}", new[] { $"{ResourceIdentifier}/{DefaultScope}" }, certificate).AccessToken);
+            var endPoint = GenericToken.GetAzureADLoginEndPoint(azureEnvironment);
+            return new GraphToken(GenericToken.AcquireApplicationToken(tenant, clientId, $"{endPoint}/{tenant}", new[] { $"{ResourceIdentifier}/{DefaultScope}" }, certificate).AccessToken);
         }
 
         /// <summary>
@@ -53,9 +49,10 @@ namespace PnP.PowerShell.Commands.Model
         /// <param name="clientId">ClientId to use to acquire the token. Required.</param>
         /// <param name="clientSecret">Client Secret to use to acquire the token. Required.</param>
         /// <returns><see cref="GraphToken"/> instance with the token</returns>
-        public static GraphToken AcquireApplicationToken(string tenant, string clientId, string clientSecret)
+        public static GraphToken AcquireApplicationToken(string tenant, string clientId, string clientSecret, AzureEnvironment azureEnvironment)
         {
-            return new GraphToken(GenericToken.AcquireApplicationToken(tenant, clientId, $"{BaseAuthority}{tenant}", new[] { $"{ResourceIdentifier}/{DefaultScope}" }, clientSecret).AccessToken);
+            var endPoint = GenericToken.GetAzureADLoginEndPoint(azureEnvironment);
+            return new GraphToken(GenericToken.AcquireApplicationToken(tenant, clientId, $"{endPoint}/{tenant}", new[] { $"{ResourceIdentifier}/{DefaultScope}" }, clientSecret).AccessToken);
         }
 
         /// <summary>
@@ -64,18 +61,19 @@ namespace PnP.PowerShell.Commands.Model
         /// <param name="clientId">ClientId to use to acquire the token. Required.</param>
         /// <param name="scopes">Array with scopes that should be requested access to. Required.</param>
         /// <returns><see cref="GraphToken"/> instance with the token</returns>
-        public static new GraphToken AcquireApplicationTokenInteractive(string clientId, string[] scopes)
+        public static new GraphToken AcquireApplicationTokenInteractive(string clientId, string[] scopes, AzureEnvironment azureEnvironment)
         {
-            return new GraphToken(GenericToken.AcquireApplicationTokenInteractive(clientId, scopes.Select(s => $"{ResourceIdentifier}/{s}").ToArray()).AccessToken);
+            return new GraphToken(GenericToken.AcquireApplicationTokenInteractive(clientId, scopes.Select(s => $"{ResourceIdentifier}/{s}").ToArray(), azureEnvironment).AccessToken);
         }
 
-        public static GraphToken AcquireApplicationTokenDeviceLogin(string clientId, string[] scopes, Action<DeviceCodeResult> callBackAction)
+        public static GraphToken AcquireApplicationTokenDeviceLogin(string clientId, string[] scopes, Action<DeviceCodeResult> callBackAction, AzureEnvironment azureEnvironment)
         {
+            var endPoint = GenericToken.GetAzureADLoginEndPoint(azureEnvironment);
             var officeManagementApiScopes = Enum.GetNames(typeof(OfficeManagementApiPermission)).Select(s => s.Replace("_", ".")).Intersect(scopes).ToArray();
             // Take the remaining scopes and try requesting them from the Microsoft Graph API
             scopes = scopes.Except(officeManagementApiScopes).ToArray();
 
-            return new GraphToken(AcquireApplicationTokenDeviceLogin(clientId, scopes, $"{BaseAuthority}organizations", callBackAction).AccessToken);
+            return new GraphToken(AcquireApplicationTokenDeviceLogin(clientId, scopes, $"{endPoint}/organizations", callBackAction).AccessToken);
         }
         /// <summary>
         /// Tries to acquire a delegated Microsoft Graph Access Token for the provided scopes using the provided credentials
@@ -85,13 +83,14 @@ namespace PnP.PowerShell.Commands.Model
         /// <param name="username">The username to authenticate with. Required.</param>
         /// <param name="securePassword">The password to authenticate with. Required.</param>
         /// <returns><see cref="GraphToken"/> instance with the token</returns>
-        public static GraphToken AcquireDelegatedTokenWithCredentials(string clientId, string[] scopes, string username, SecureString securePassword)
+        public static GraphToken AcquireDelegatedTokenWithCredentials(string clientId, string[] scopes, string username, SecureString securePassword, AzureEnvironment azureEnvironment)
         {
+            var endPoint = GenericToken.GetAzureADLoginEndPoint(azureEnvironment);
             var officeManagementApiScopes = Enum.GetNames(typeof(OfficeManagementApiPermission)).Select(s => s.Replace("_", ".")).Intersect(scopes).ToArray();
             // Take the remaining scopes and try requesting them from the Microsoft Graph API
             scopes = scopes.Except(officeManagementApiScopes).ToArray();
-            
-            return new GraphToken(AcquireDelegatedTokenWithCredentials(clientId, scopes.Select(s => $"{ResourceIdentifier}/{s}").ToArray(), $"{BaseAuthority}organizations/", username, securePassword).AccessToken);
+
+            return new GraphToken(AcquireDelegatedTokenWithCredentials(clientId, scopes.Select(s => $"{ResourceIdentifier}/{s}").ToArray(), $"{endPoint}/organizations/", username, securePassword).AccessToken);
         }
     }
 }