+++ Commands/Provisioning/Site/ApplyProvisioningTemplate.cs
@@ -369,6 +369,11 @@ PS:> Apply-PnPProvisioningTemplate -Path NewTemplate.xml -ExtensibilityHandlers 
 #if !ONPREMISES
             using (var provisioningContext = new PnPProvisioningContext(async (resource, scope) =>
             {
+                if (resource.ToLower().StartsWith("https://"))
+                {
+                    var uri = new Uri(resource);
+                    resource = uri.Authority;
+                }
                 // Get Azure AD Token
                 if (PnPConnection.CurrentConnection != null)
                 {
@@ -377,7 +382,7 @@ PS:> Apply-PnPProvisioningTemplate -Path NewTemplate.xml -ExtensibilityHandlers 
                     {
                         // Authenticated using -Graph or using another way to retrieve the accesstoken with Connect-PnPOnline
                         return await Task.FromResult(graphAccessToken);
-                        
+
                     }
                 }
 
@@ -389,7 +394,7 @@ PS:> Apply-PnPProvisioningTemplate -Path NewTemplate.xml -ExtensibilityHandlers 
                 else
                 {
                     // No token...
-                    return await Task.FromResult<string>(null);
+                    throw new PSInvalidOperationException("Your template contains artifacts that require an access token. Please provide consent to the PnP Management Shell application first by executing: Connect-PnPOnline -Graph -LaunchBrowser");
                 }
             }))
             {